Fixing bugs: changed Token and Error to have separate returns because of LoginPage being unable to differentiate between the two
